Update spdlog to 1.14.1

+++ src/gui/qt/main.cpp
@@ -58,7 +58,7 @@ void logRuntimeEnvironment() {
   const auto logger = loot::getLogger();
   if (logger) {
     logger->info("Running {}-bit LOOT on {} and the {} CPU architecture",
-                 QSysInfo::WordSize,
+                 static_cast<int>(QSysInfo::WordSize),
                  QSysInfo::prettyProductName().toStdString(),
                  QSysInfo::currentCpuArchitecture().toStdString());
     logger->info("LOOT version: {}+{}",
